Extract `cvd/fetch/target_directories.{cc,h}`
Bug: b/424021019
Test: bazel test '//...' --test_summary=terse
