+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -184,7 +184,7 @@ cc_library(
         "//cuttlefish/host/commands/cvd/fetch:get_optional",
         "//cuttlefish/host/commands/cvd/fetch:host_package",
         "//cuttlefish/host/commands/cvd/fetch:substitute",
-        "//cuttlefish/host/commands/cvd/fetch:vector_flags",
+        "//cuttlefish/host/commands/cvd/fetch:target_directories",
         "//cuttlefish/host/libs/config:fetcher_config",
         "//cuttlefish/host/libs/image_aggregator",
         "//cuttlefish/host/libs/web/http_client:curl_global_init",
@@ -322,6 +322,22 @@ cc_library(
 clang_tidy_test(
     name = "substitute_clang_tidy",
     srcs = [":substitute"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
+    name = "target_directories",
+    srcs = ["target_directories.cc"],
+    hdrs = ["target_directories.h"],
+    copts = COPTS,
+    deps = [
+        "//cuttlefish/host/commands/cvd/fetch:get_optional",
+    ],
+)
+
+clang_tidy_test(
+    name = "target_directories_clang_tidy",
+    srcs = [":target_directories"],
     tags = ["clang_tidy", "clang-tidy"],
 )
