+++ base/cvd/cuttlefish/host/commands/cvd/fetch/target_directories.h
@@ -1,0 +1,36 @@
+//
+// Copyright (C) 2019 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+namespace cuttlefish {
+
+struct TargetDirectories {
+  static TargetDirectories Create(
+      const std::string& target_directory,
+      const std::vector<std::string>& target_subdirectories, int index,
+      bool append_subdirectory);
+
+  std::string root;
+  std::string otatools;
+  std::string default_target_files;
+  std::string system_target_files;
+  std::string chrome_os;
+};
+
+}  // namespace cuttlefish
